chore(select-input): move to packages
