+++ packages/components/inputs/select-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './select-input';
+
+export { default as version } from './version';
