+++ packages/components/inputs/select-input/src/select-input.js
@@ -16,8 +16,8 @@ import {
   SafeHTMLElement,
   filterDataAttributes,
 } from '@commercetools-uikit/utils';
-import createSelectStyles from '../../internals/create-select-styles';
-import messages from '../../internals/messages/select';
+import createSelectStyles from '../../../../../src/components/internals/create-select-styles';
+import messages from '../../../../../src/components/internals/messages/select';
 
 const useTheme = () => useContext(ThemeContext);
