--- src/components/inputs/select-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './select-input';
